requirements.txt
asgiref==3.2.3
backcall==0.1.0
decorator==4.4.2
Django==3.0.7
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
et-xmlfile==1.1.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
numpy==1.18.5
openpyxl==3.0.7
pandas==1.0.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
six==1.15.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.9
xlrd==2.0.1
